+++ src/Model/Reports/CybersecurityReportFormData.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Model\Reports;
+
+use App\Entity\DataProvider;
+
+class CybersecurityReportFormData
+{
+    public DataProvider $dataProvider;
+    public string $versionTitle;
+    public \DateTimeInterface $fromDate;
+    public \DateTimeInterface $toDate;
+}